添加在添加新鲜事图片后的待审核图片

+++ fresh-system/src/main/java/com/creamakers/fresh/system/domain/vo/response/FreshNewsDetailResp.java
@@ -35,7 +35,7 @@ public class FreshNewsDetailResp {
     private List<String> tags;
 
     // 点赞数量
-    private Integer liked;
+    private Long liked;
 
     // 评论数量
     private Integer comments;
